cmd/kubectl-gadget: Add undeploy command
Inspektor Gadget doesn't provide a proper way to remove it from the
cluster. Currently we suggest to use "kubectl gadget deploy | kubectl
delete -f -". This method has some problems now that we have a custom
trace resource as it doesn't remove the traces but only the static
resources that were created at the beginning. In order to perform a
proper cleanup of all Inspektor Gadget related resources this commit
implements a "undeploy" command that talks directly to the API server.

Traces are a bit tricky to remove. This logic tries to remove them
before removing the DaemonSet to give an opportunity to the trace
controller for removing them. If the removal doesn't succeeds within a
given period of time, we remove the finalizer to let Kubernetes remove
them without waiting for the trace controller.